+++ src/base/Timer.h
@@ -41,26 +41,25 @@ public:
 			nullptr) noexcept;
 	//Generate a 64-bit seed from current time for seeding the RNGs
 	static unsigned long long timeSeed() noexcept;
-
-	//Creates a timer file descriptor
-	static int openTimerfd(bool blocking = false);
-	/*
-	 * Sets expiration and periodic interval in milliseconds.
-	 * Setting <expiration> to zero(0) will disable the timer.
-	 */
-	static void setTimerfd(int fd, unsigned int expiration,
-			unsigned int interval);
-	//Returns the settings of a timer file descriptor in milliseconds
-	static void getTimerfdSettings(int fd, unsigned int &expiration,
-			unsigned int &interval);
-	//Closes a timer file descriptor
-	static int closeTimerfd(int fd) noexcept;
 private:
 	//Wrapper for clock_gettime, returns current time
 	static unsigned long long currentTime() noexcept;
 	//Returns the difference of the microsecond precision values in seconds
 	static double difference(unsigned long long start,
 			unsigned long long end) noexcept;
+public:
+	//Nanoseconds in a Second
+	static constexpr long NS_IN_SEC = 1000000000L;
+	//Microseconds in a Second
+	static constexpr long MS_IN_SEC = 1000000L;
+	//Milliseconds in a Second
+	static constexpr long MILS_IN_SEC = 1000L;
+	//Nanoseconds in a Microsecond
+	static constexpr long NS_IN_MS = (NS_IN_SEC / MS_IN_SEC);
+	//Nanoseconds in a Millisecond
+	static constexpr long NS_IN_MILS = (NS_IN_SEC / MILS_IN_SEC);
+	//Microseconds in a Millisecond
+	static constexpr long MS_IN_MILS = (MS_IN_SEC / MILS_IN_SEC);
 private:
 	//Time in microseconds, sufficient for 300,000 years
 	unsigned long long t;