+++ docs/ChangeLog.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Signal::sigchildHandler** should preserve the errno.
 - Data types of **Selector** event and flag enumerations should be 32-bit unsigned int.
 
+### Removed
+
+- All the **timerfd** related methods from the **Timer** class.
+
 ## [4.0.0] - 2021-04-01
 
 ### Added